updateclasses

+++ Real_Estate/src/main/java/com/pms/RealEstate/model/Buying.java
@@ -7,30 +7,36 @@ public class Buying {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long buying_id;
+    private int buying_id;
 
     @OneToOne
     @JoinColumn(name = "property_id")
-    private Property property;
+    private Property property;    
+    
 
-    private Double expected_rate;
+    private double expected_rate;
 
 	public Buying() {
 		super();
 	}
 
-	public Buying(Long buying_id, Property property, Double expected_rate) {
+	public Buying(int buying_id, Property property, double expected_rate) {
 		super();
 		this.buying_id = buying_id;
 		this.property = property;
 		this.expected_rate = expected_rate;
 	}
 
-	public Long getBuying_id() {
+	
+	
+	
+	
+	
+	public int getBuying_id() {
 		return buying_id;
 	}
 
-	public void setBuying_id(Long buying_id) {
+	public void setBuying_id(int buying_id) {
 		this.buying_id = buying_id;
 	}
 
@@ -42,11 +48,11 @@ public class Buying {
 		this.property = property;
 	}
 
-	public Double getExpected_rate() {
+	public double getExpected_rate() {
 		return expected_rate;
 	}
 
-	public void setExpected_rate(Double expected_rate) {
+	public void setExpected_rate(double expected_rate) {
 		this.expected_rate = expected_rate;
 	}
 
@@ -55,5 +61,4 @@ public class Buying {
 		return "Buying [buying_id=" + buying_id + ", property=" + property + ", expected_rate=" + expected_rate + "]";
 	}
 
-     
 }
